+++ pkg/depthmonitor/interface.go
@@ -1,0 +1,30 @@
+// Copyright 2022 The Swarm Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package depthmonitor
+
+import "github.com/ethersphere/bee/pkg/topology"
+
+// ReserveReporter interface defines the functionality required from the local storage
+// of the node to report information about the reserve. The reserve storage is the storage
+// pledged by the node to the network.
+type ReserveReporter interface {
+	// Current size of the reserve.
+	Size() (uint64, error)
+	// Capacity of the reserve that is configured.
+	Capacity() uint64
+}
+
+// SyncReporter interface needs to be implemented by the syncing component of the node (pullsync).
+type SyncReporter interface {
+	// Rate of syncing in terms of chunks/sec.
+	Rate() float64
+}
+
+// Topology interface encapsulates the functionality required by the topology component
+// of the node.
+type Topology interface {
+	topology.NeighborhoodDepther
+	topology.SetStorageDepther
+}
